Update nominatim.js
 Copilot Autofix AI generated 

To fix the problem, we need to escape the . character in the regular expression to ensure it matches only the literal . character and not any character. This can be done by replacing . with \\. in the regular expression. This change should be made in the file test/spec/services/nominatim.js on line 118.