Update wikipedia.js
 Copilot Autofix AI 

To fix the problem, we need to ensure that the dot character in the regular expression is correctly escaped to match a literal dot. This can be achieved by using double backslashes (\\.) in the string literal, which will be interpreted as a single backslash in the regular expression, thus correctly escaping the dot character.